backend/src/requirements.txt
flask
flask[async]
flask-restful
better-profanity
openai
python-dotenv
